add news publish

+++ conf/evolutions/default/4.sql
@@ -1,11 +1,11 @@
 # --- !Ups
 
 CREATE TABLE newsRead (
-  id   SERIAL PRIMARY KEY,
-  k_id   INT NOT NULL,
-  parent_id   INT NOT NULL,
-  news_id   INT,
-  readTime DATE         NOT NULL
+  id        SERIAL PRIMARY KEY,
+  k_id      LONG NOT NULL,
+  parent_id LONG NOT NULL,
+  news_id   LONG,
+  readTime  DATE NOT NULL
 );
 
 # --- !Downs
